Refactor filesystem handling and remove unused features
- Updated Cargo.toml to streamline filesystem dependencies by removing redundant entries and introducing new features for ext4 and fat filesystems.
- Modified board configuration files to eliminate the "fs" feature where it was no longer necessary.
- Adjusted tests and build scripts to reflect the removal of the "fs" feature, ensuring compatibility with the new filesystem structure.
- Introduced a new `fs_policy.rs` file to manage filesystem detection and initialization, enhancing the overall organization and clarity of the filesystem handling logic.
