+++ os/arceos/modules/axruntime/Cargo.toml
@@ -25,14 +25,9 @@ tls = ["ax-hal/tls", "ax-task?/tls"]
 plat-dyn = ["ax-hal/plat-dyn", "paging", "buddy-slab"]
 
 display = ["dep:ax-display", "ax-driver/display"]
-fs = [
-    "dep:ax-errno",
-    "dep:ax-fs",
-    "dep:rd-block",
-    "dep:spin",
-    "dep:axklib",
-    "ax-driver/block",
-]
+fs = ["dep:ax-errno", "dep:ax-fs", "ax-driver/block"]
+fs-ext4 = ["fs", "ax-fs/ext4"]
+fs-fat = ["fs", "ax-fs/fat"]
 input = ["dep:ax-input", "ax-driver/input"]
 net = ["dep:ax-net", "ax-driver/net"]
 net-ng = ["dep:ax-net-ng", "dep:rd-net", "dep:spin", "dep:axklib", "ax-driver/net"]
@@ -65,7 +60,6 @@ ax-crate-interface = { workspace = true }
 ax-ctor-bare = { workspace = true }
 indoc = "2"
 ax-percpu = { workspace = true, optional = true }
-rd-block = { workspace = true, optional = true }
 rd-net = { workspace = true, optional = true }
 rdrive.workspace = true
 spin = { workspace = true, optional = true }
