+++ os/axvisor/configs/board/orangepi-5-plus.toml
@@ -2,7 +2,6 @@ env = { AX_IP = "10.0.2.15", AX_GW = "10.0.2.2" }
 features = [
   "sdmmc",
   "rockchip-soc",
-  "fs",
 ]
 log = "Info"
 plat_dyn = true
