+++ os/axvisor/configs/board/qemu-loongarch64.toml
@@ -4,7 +4,6 @@ features = [
   "ax-driver/virtio-blk",
   "ax-driver/pci",
   "ept-level-4",
-  "fs",
 ]
 log = "Info"
 max_cpu_num = 1
