+++ os/axvisor/configs/board/qemu-x86_64.toml
@@ -4,7 +4,6 @@ features = [
     "ax-driver/virtio-blk",
     "ax-driver/pci",
     "ept-level-4",
-    "fs",
 ]
 log = "Info"
 target = "x86_64-unknown-none"
