+++ os/axvisor/configs/board/roc-rk3568-pc.toml
@@ -1,6 +1,5 @@
 env = { AX_IP = "10.0.2.15", AX_GW = "10.0.2.2" }
 features = [
-    "fs",
     "sdmmc",
 ]
 log = "Info"
