+++ scripts/axbuild/src/axvisor/board.rs
@@ -98,7 +98,7 @@ mod tests {
             r#"
 env = { AX_IP = "10.0.2.15", AX_GW = "10.0.2.2" }
 target = "aarch64-unknown-none-softfloat"
-features = ["fs"]
+features = []
 log = "Info"
 "#,
         );
@@ -128,7 +128,7 @@ log = "Info"
             r#"
 env = { AX_IP = "10.0.2.15", AX_GW = "10.0.2.2" }
 target = "aarch64-unknown-none-softfloat"
-features = ["fs", "ax-driver/rockchip-sdhci"]
+features = ["ax-driver/rockchip-sdhci"]
 log = "Info"
 plat_dyn = true
 "#,
