+++ scripts/repo/repos.csv
@@ -47,7 +47,6 @@ https://github.com/arceos-org/riscv_plic,,drivers/intc/riscv_plic,ArceOS,
 https://github.com/rcore-os/bitmap-allocator,,components/bitmap-allocator,rCore,
 https://github.com/Starry-OS/axbacktrace,,components/axbacktrace,Starry,
 https://github.com/Starry-OS/axpoll,,components/axpoll,Starry,
-https://github.com/Starry-OS/axfs-ng-vfs,,components/axfs-ng-vfs,Starry,
 https://github.com/Starry-OS/rsext4,,components/rsext4,Starry,
 https://github.com/Starry-OS/scope-local,,components/scope-local,Starry,
 https://github.com/Starry-OS/starry-process,,components/starry-process,Starry,
